decouple to AbstractRector
