+++ src/Rector/Contrib/Nette/DI/CompilerArgumentsRector.php
@@ -5,12 +5,8 @@ namespace Rector\Rector\Contrib\Nette\DI;
 use PhpParser\Node;
 use PhpParser\Node\Arg;
 use PhpParser\Node\Expr\MethodCall;
-use PhpParser\Node\Stmt\Expression;
-use Rector\Builder\StatementGlue;
-use Rector\Node\Attribute;
 use Rector\NodeAnalyzer\MethodCallAnalyzer;
 use Rector\Rector\AbstractRector;
-use SplObjectStorage;
 
 /**
  * Nette\DI\Compiler::compile arguments are deprecated, use Compiler::addConfig() and Compiler::setClassName().
@@ -30,21 +26,9 @@ final class CompilerArgumentsRector extends AbstractRector
      */
     private $methodCallAnalyzer;
 
-    /**
-     * @var SplObjectStorage|Expression[]
-     */
-    private $expressionsToPrepend = [];
-
-    /**
-     * @var StatementGlue
-     */
-    private $statementGlue;
-
-    public function __construct(MethodCallAnalyzer $methodCallAnalyzer, StatementGlue $statementGlue)
+    public function __construct(MethodCallAnalyzer $methodCallAnalyzer)
     {
-        $this->expressionsToPrepend = new SplObjectStorage;
         $this->methodCallAnalyzer = $methodCallAnalyzer;
-        $this->statementGlue = $statementGlue;
     }
 
     public function isCandidate(Node $node): bool
@@ -58,30 +42,18 @@ final class CompilerArgumentsRector extends AbstractRector
     }
 
     /**
-     * @todo build into abstract?
-     *
-     * @param Node[] $nodes
-     * @return Node[]
-     */
-    public function afterTraverse(array $nodes): array
-    {
-        return $this->processNodes($nodes);
-    }
-
-    /**
      * @param MethodCall $methodCallNode
      */
     public function refactor(Node $methodCallNode): ?Node
     {
         $oldArguments = $methodCallNode->args;
-        $nodesToPrepend = [];
 
         $addConfigMethodCallNode = $this->cloneMethodWithNameAndArgument(
             $methodCallNode,
             'addConfig',
             $oldArguments[0]
         );
-        $nodesToPrepend[] = new Expression($addConfigMethodCallNode);
+        $this->prependNodeBehindNode($addConfigMethodCallNode, $methodCallNode);
 
         if (isset($oldArguments[1])) {
             $setClassNameMethodCallNode = $this->cloneMethodWithNameAndArgument(
@@ -89,37 +61,12 @@ final class CompilerArgumentsRector extends AbstractRector
                 'setClassName',
                 $oldArguments[1]
             );
-            $nodesToPrepend[] = new Expression($setClassNameMethodCallNode);
+            $this->prependNodeBehindNode($setClassNameMethodCallNode, $methodCallNode);
         }
-
-        $parentExpressionNode = $methodCallNode->getAttribute(Attribute::PARENT_NODE);
-        $this->expressionsToPrepend[$parentExpressionNode] = $nodesToPrepend;
 
         $methodCallNode->args = [];
 
         return $methodCallNode;
-    }
-
-    /**
-     * @param Node[] $nodes
-     * @return Node[] array
-     */
-    private function processNodes(array $nodes): array
-    {
-        foreach ($nodes as $i => $node) {
-            if ($node instanceof Expression) {
-                if (! isset($this->expressionsToPrepend[$node])) {
-                    continue;
-                }
-
-                $nodes = $this->statementGlue->insertNewNodesAfter($nodes, $this->expressionsToPrepend[$node], $i);
-                $this->expressionsToPrepend[$node] = null;
-            } elseif (isset($node->stmts)) {
-                $node->stmts = $this->processNodes($node->stmts);
-            }
-        }
-
-        return $nodes;
     }
 
     private function cloneMethodWithNameAndArgument(
